Use longs instead of ints

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day18/AddInstruction.java
@@ -21,7 +21,7 @@ class AddInstruction extends UpdateRegisterInstruction {
     }
     
     @Override
-    int computeNewValue(int oldValue, State state) {
+    long computeNewValue(long oldValue, State state) {
         return oldValue + getValue(state);
     }
