+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day18/SetInstruction.java
@@ -21,7 +21,7 @@ class SetInstruction extends UpdateRegisterInstruction {
     }
     
     @Override
-    int computeNewValue(int oldValue, State state) {
+    long computeNewValue(long oldValue, State state) {
         return getValue(state);
     }
